showModal now checks for connectorSupportsUsb value
Since not all connectors will support USB fallback for the Ledger button
I've added a connectorSupportsUsb property to the showModal input object
which determines if we care about USB support. Now we only show the
platform not supported modal due to lack of USB support if this property
is true.