Increment version to fix release concurrency problem
The latest Connect Kit release was successfuly published to NPM with
version 1.0.0-beta.6 but the version increase was not commited to the repo
because the Connect Kit loader release action that was running at the
same time had already commited. This should not happen again since the
same concurrency tag is used for both now.